+++ assets/system/game_modules/sphere-runtime.js
@@ -30,6 +30,12 @@
  *  POSSIBILITY OF SUCH DAMAGE.
 **/
 
+export default
+class RT
+{
+	static get Version() { return 0; }
+}
+
 export { default as Console } from 'console';
 export { default as FocusTarget } from 'focus-target';
 export { default as Music } from 'music';