+++ CHANGELOG.md
@@ -4,6 +4,7 @@ neoSphere Changelog
 Since Last Release
 ------------------
 
+* Adds `RT.Version` for getting the current API revision of the Sphere Runtime.
 * Disables `new Surface(fileName)` for games targeting API 4 or higher.